pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-intelligence"
version = "0.1.0"
description = "Modular stock intelligence platform for Indian equities"
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENCE" }
authors = [{ name = "HotShot0901" }]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "yfinance>=0.2.50",
    "kiteconnect>=5.0",
    "pandas>=2.2",
    "numpy>=2.0",
    "apscheduler>=3.10",
    "jinja2>=3.1",
    "openai>=1.55",
    "textblob>=0.18",
    "feedparser>=6.0",
    "rapidfuzz>=3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "hypothesis>=6.115",
    "ruff>=0.8",
    "mypy>=1.13",
    "types-pyyaml",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]
# N812: date aliased to DateType to avoid shadowing Pydantic `date` fields.
# UP046/UP047: PEP 695 generic syntax — keep explicit TypeVar for Pydantic+mypy compatibility.
# B008: FastAPI's Depends()/Query()/Body() idiom requires function calls in argument defaults.
ignore = ["E501", "N812", "UP046", "UP047", "B008"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true

[[tool.mypy.overrides]]
module = [
    "yfinance",
    "apscheduler.*",
    "kiteconnect",
    "kiteconnect.*",
    "textblob",
    "feedparser",
    "rapidfuzz",
]
ignore_missing_imports = true
